Further readme edits
So, I got as far as:
Create an API key and service account named api-tester
I followed the directions, but I think you are assuming that you follow the readme.md of the uber project in order to get a project setup and enabled with the API's. I came unstuck as I went into the Google pages and ended up creating a new project in order to enable api-tester.
There is also great dependency on the creation and setting of the shell var from the uber project readme.md but it's not mentioned in this project.

So my concern is that this is not called out. It's like "Don't start here, start one level above, get you env stetup and then go to here" but will that scale to multiple projects? Surely you need one project per set of telemetry and you should not add all to the same project you set up from the uber readme.md. Or am I wrong?